Improvement for image preparation and fewer arguments
If neither `SOURCE_RPM` nor `SPEC_FILE` is specified, `mock init` is
run, and at the end an option is given to checkpoint the image
using `docker commit`. This makes it easier to create an image
with mock already pre-configured, which makes future runs much
faster.

The `SOURCES` argument is now optional, `SPEC_FILE` is sufficient
to start the build.